Made test pass

+++ system_tests/test_dr.py
@@ -199,7 +199,8 @@ class TestDr:
 
     @pytest.fixture
     def simulation(self) -> subprocess.Popen:
-        shutil.rmtree(config.SIMULATION_DIRECTORY_PATH)
+        if os.path.exists(config.SIMULATION_DIRECTORY_PATH):
+            shutil.rmtree(config.SIMULATION_DIRECTORY_PATH)
         os.makedirs(config.SIMULATION_DIRECTORY_PATH)
         shutil.copy2(config.MAV_PARAM_FILE_PATH, config.SIMULATION_DIRECTORY_PATH)
         sitl_cmd = [
@@ -227,7 +228,9 @@ class TestDr:
 
     @pytest.fixture
     def drone(self, simulation: subprocess.Popen) -> drone_controller.Drone:
-        my_drone = drone_controller.Drone(source_system=config.MISSION_COMPUTER_MAVLINK_SYSTEM_ID)
+        my_drone = drone_controller.Drone(
+            source_system=config.MISSION_COMPUTER_MAVLINK_SYSTEM_ID
+        )
         my_drone.connect(
             f"udp:localhost:{config.TESTS_PORT}",
             connection_timeout_sec=config.DRONE_CONNECT_TO_SIMULATION_TIMEOUT_SEC,
@@ -293,7 +296,10 @@ class TestDr:
     #     [config.DR_AXIS_DISTANCE_DEGREES, -config.DR_AXIS_DISTANCE_DEGREES, 0],
     # )
     # Use this and comment the 2 above markers for a single test:
-    @pytest.mark.parametrize("offset_latitude, offset_longitude", [(config.DR_AXIS_DISTANCE_DEGREES, config.DR_AXIS_DISTANCE_DEGREES)])
+    @pytest.mark.parametrize(
+        "offset_latitude, offset_longitude",
+        [(config.DR_AXIS_DISTANCE_DEGREES, config.DR_AXIS_DISTANCE_DEGREES)],
+    )
     def test_dr_direction(
         self,
         flying_drone: drone_controller.Drone,
@@ -320,8 +326,9 @@ class TestDr:
                 flying_drone.mode == "GUIDED"
             ), "Something wrong happend, the drone is not in GUIDED mode"
             # Let the autopilot get the drone position before we disable it.
-            utils.wait(2)
+            utils.wait(10)
             flying_drone.sim_gps_disable = True
+            utils.wait(10, absolute=True)
             self.wait_for_mode_change(
                 drone_obj=flying_drone,
                 mode="ALT_HOLD",
@@ -405,17 +412,17 @@ class TestDr:
     def test_dr_trigger_rc_failure_and_gps_failure_imidiatly_after(self):
         """Should change mode to LAND and then cancel it and trigger DR"""
         assert False
-    
+
     @pytest.mark.system
     @pytest.mark.skip(reason="WIP")
     def test_dr_gps_returns_no_rc():
         assert False
-    
+
     @pytest.mark.system
     @pytest.mark.skip(reason="WIP")
     def test_dr_gps_returns_rc_healthy():
         assert False
-    
+
     @pytest.mark.system
     @pytest.mark.skip(reason="WIP")
     def test_dr_rc_returns_no_gps():
